backend/requirements.txt
django==3.2.14
environs[django]==9.3.5
psycopg2-binary==2.9.2
djangorestframework==3.13.1
django-extensions==3.2.0
Pillow==9.2.0
django-taggit==3.0.0
